+++ CHANGELOG.md
@@ -12,3 +12,9 @@
 * scripts/*.vcf            : added the FORMAT/AD tag
 
 * scripts/filter2.sh       : added new script to filter the original mutect2.mutect2 calls => mutect2.mutect2.00.orig
+
+# 2024/07/31 #
+
+
+* scripts/filter2.sh	   : file path update
+